C program that prints the size of various types

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int n;
 	float c;
 
-	printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(p));
-	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(r));
-	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(n));
-	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a char: %lu bytes(s)\n", sizeof(p));
+	printf("Size of an int: %lu bytes(s)\n", sizeof(r));
+	printf("Size of a long int: %lu bytes(s)\n", sizeof(i));
+	printf("Size of a long long int: %lu bytes(s)\n", sizeof(n));
+	printf("Size of a float: %lu bytes(s)\n", sizeof(c));
 	return (0);
 }
